refactored today widget to mvvm proposed architecture + added accessibility hints
