+++ Extensions/Today/ButtonWithSublabel.swift
@@ -1,0 +1,60 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import UIKit
+
+class ButtonWithSublabel: UIButton {
+    lazy var subtitleLabel = UILabel()
+
+     lazy var label = UILabel()
+
+     required init?(coder aDecoder: NSCoder) {
+         fatalError("init(coder:) has not been implemented")
+     }
+
+     convenience init() {
+         self.init(frame: .zero)
+     }
+
+     override init(frame: CGRect) {
+         super.init(frame: frame)
+         performLayout()
+     }
+
+     fileprivate func performLayout() {
+         let titleLabel = self.label
+         self.titleLabel?.removeFromSuperview()
+         addSubview(titleLabel)
+
+         let imageView = self.imageView!
+         let subtitleLabel = self.subtitleLabel
+         self.addSubview(subtitleLabel)
+
+         imageView.snp.makeConstraints { make in
+             make.centerY.left.equalTo(10)
+             make.width.equalTo(TodayUX.copyLinkImageWidth)
+         }
+
+         titleLabel.snp.makeConstraints { make in
+             make.left.equalTo(imageView.snp.right).offset(10)
+             make.trailing.top.equalTo(self)
+             make.height.equalTo(15)
+         }
+        titleLabel.numberOfLines = 1
+        titleLabel.lineBreakMode = .byWordWrapping
+
+         subtitleLabel.lineBreakMode = .byTruncatingTail
+         subtitleLabel.snp.makeConstraints { make in
+             make.bottom.equalTo(self).inset(10)
+             make.top.equalTo(titleLabel.snp.bottom).offset(3)
+             make.leading.trailing.equalTo(titleLabel)
+             make.height.equalTo(10)
+         }
+     }
+
+     override func setTitle(_ text: String?, for state: UIControl.State) {
+         self.label.text = text
+         super.setTitle(text, for: state)
+     }
+}
