+++ Extensions/Today/TodayViewController.swift
@@ -10,40 +10,11 @@ import XCGLogger
 
 private let log = Logger.browserLogger
 
-struct TodayStrings {
-    static let NewPrivateTabButtonLabel = NSLocalizedString("TodayWidget.NewPrivateTabButtonLabel", tableName: "Today", value: "Private Search", comment: "New Private Tab button label")
-    static let NewTabButtonLabel = NSLocalizedString("TodayWidget.NewTabButtonLabel", tableName: "Today", value: "New Search", comment: "New Tab button label")
-    static let GoToCopiedLinkLabel = NSLocalizedString("TodayWidget.GoToCopiedLinkLabel", tableName: "Today", value: "Go to copied link", comment: "Go to link on clipboard")
-}
-
-private struct TodayUX {
-    static let backgroundHightlightColor = UIColor(white: 216.0/255.0, alpha: 44.0/255.0)
-    static let linkTextSize: CGFloat = 9.0
-    static let labelTextSize: CGFloat = 12.0
-    static let imageButtonTextSize: CGFloat = 13.0
-    static let copyLinkImageWidth: CGFloat = 20
-    static let margin: CGFloat = 8
-    static let buttonsHorizontalMarginPercentage: CGFloat = 0.1
-    static let buttonStackViewSpacing: CGFloat = 30.0
-    static var labelColor: UIColor {
-        if #available(iOS 13, *) {
-            return UIColor(named: "widgetLabelColors") ?? UIColor(rgb: 0x242327)
-        } else {
-            return UIColor(rgb: 0x242327)
-        }
-    }
-    static var subtitleLabelColor: UIColor {
-        if #available(iOS 13, *) {
-            return UIColor(named: "subtitleLableColor") ?? UIColor(rgb: 0x38383C)
-        } else {
-            return UIColor(rgb: 0x38383C)
-        }
-    }
-}
-
 @objc (TodayViewController)
-class TodayViewController: UIViewController, NCWidgetProviding {
-    var copiedURL: URL?
+class TodayViewController: UIViewController, NCWidgetProviding, TodayWidgetAppearanceDelegate {
+
+    let viewModel = TodayWidgetViewModel()
+    let model = TodayModel()
 
     fileprivate lazy var newTabButton: ImageButtonWithLabel = {
         let imageButton = ImageButtonWithLabel()
@@ -51,10 +22,11 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         imageButton.label.text = TodayStrings.NewTabButtonLabel
         let button = imageButton.button
         button.setImage(UIImage(named: "search-button")?.withRenderingMode(.alwaysOriginal), for: .normal)
+        button.accessibilityLabel = TodayStrings.NewTabButtonLabel
         let label = imageButton.label
         label.textColor = TodayUX.labelColor
         label.tintColor = TodayUX.labelColor
-        label.font = UIFont.systemFont(ofSize: TodayUX.imageButtonTextSize)
+        label.font = UIFontMetrics(forTextStyle: .body).scaledFont(for: UIFont.preferredFont(forTextStyle: .body).withSize(TodayUX.imageButtonTextSize))
         imageButton.sizeToFit()
         return imageButton
     }()
@@ -65,10 +37,11 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         imageButton.label.text = TodayStrings.NewPrivateTabButtonLabel
         let button = imageButton.button
         button.setImage(UIImage(named: "private-search")?.withRenderingMode(.alwaysOriginal), for: .normal)
+        button.accessibilityLabel = TodayStrings.NewPrivateTabButtonLabel
         let label = imageButton.label
         label.textColor = TodayUX.labelColor
         label.tintColor = TodayUX.labelColor
-        label.font = UIFont.systemFont(ofSize: TodayUX.imageButtonTextSize)
+        label.font = UIFontMetrics(forTextStyle: .body).scaledFont(for: UIFont.preferredFont(forTextStyle: .body).withSize(TodayUX.imageButtonTextSize))
         imageButton.sizeToFit()
         return imageButton
     }()
@@ -81,12 +54,14 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         button.setBackgroundColor(UIColor.clear, forState: .normal)
         button.setBackgroundColor(TodayUX.backgroundHightlightColor, forState: .highlighted)
         button.setImage(UIImage(named: "copy_link_icon")?.withRenderingMode(.alwaysOriginal), for: .normal)
-        button.label.font = UIFont.systemFont(ofSize: TodayUX.labelTextSize)
-        button.subtitleLabel.font = UIFont.systemFont(ofSize: TodayUX.linkTextSize)
+        button.label.font = UIFontMetrics(forTextStyle: .body).scaledFont(for: UIFont.preferredFont(forTextStyle: .body).withSize(TodayUX.labelTextSize))
+        button.accessibilityLabel = TodayStrings.GoToCopiedLinkLabel
         button.label.textColor = TodayUX.labelColor
         button.label.tintColor = TodayUX.labelColor
+        button.label.sizeToFit()
         button.subtitleLabel.textColor = TodayUX.subtitleLabelColor
         button.subtitleLabel.tintColor = TodayUX.subtitleLabelColor
+        button.subtitleLabel.font = UIFontMetrics(forTextStyle: .body).scaledFont(for: UIFont.preferredFont(forTextStyle: .body).withSize(TodayUX.linkTextSize))
         return button
     }()
 
@@ -108,16 +83,9 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         return stackView
     }()
 
-    fileprivate var scheme: String {
-        guard let string = Bundle.main.object(forInfoDictionaryKey: "MozInternalURLScheme") as? String else {
-            // Something went wrong/weird, but we should fallback to the public one.
-            return "firefox"
-        }
-        return string
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
+        viewModel.setViewDelegate(todayViewDelegate: self)
         let widgetView: UIView!
         self.extensionContext?.widgetLargestAvailableDisplayMode = .compact
 
@@ -148,7 +116,7 @@ class TodayViewController: UIViewController, NCWidgetProviding {
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        updateCopiedLink()
+        viewModel.updateCopiedLink()
     }
 
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
@@ -160,18 +128,14 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         return .zero
     }
 
-    func updateCopiedLink() {
-        UIPasteboard.general.asyncURL().uponQueue(.main) { res in
-            if let copiedURL: URL? = res.successValue,
-                let url = copiedURL {
-                self.openCopiedLinkButton.isHidden = false
-                self.openCopiedLinkButton.subtitleLabel.isHidden = SystemUtils.isDeviceLocked()
-                self.openCopiedLinkButton.subtitleLabel.text = url.absoluteDisplayString
-                self.copiedURL = url
-            } else {
-                self.openCopiedLinkButton.isHidden = true
-                self.copiedURL = nil
-            }
+    func updateCopiedLinkInView(clipboardURL: URL?) {
+        if let url = clipboardURL {
+            self.openCopiedLinkButton.isHidden = false
+            self.openCopiedLinkButton.subtitleLabel.isHidden = SystemUtils.isDeviceLocked()
+            self.openCopiedLinkButton.subtitleLabel.text = url.absoluteDisplayString
+        } else {
+            self.openCopiedLinkButton.isHidden = true
+            self.openCopiedLinkButton.subtitleLabel.isHidden = SystemUtils.isDeviceLocked()
         }
     }
 
@@ -183,129 +147,18 @@ class TodayViewController: UIViewController, NCWidgetProviding {
     @objc func onPressNewPrivateTab(_ view: UIView) {
         openContainingApp("?private=true")
     }
-
+    //TODO: Move it to model??
     fileprivate func openContainingApp(_ urlSuffix: String = "") {
-        let urlString = "\(scheme)://open-url\(urlSuffix)"
+        let urlString = "\(model.scheme)://open-url\(urlSuffix)"
         self.extensionContext?.open(URL(string: urlString)!) { success in
             log.info("Extension opened containing app: \(success)")
         }
     }
 
     @objc func onPressOpenClibpoard(_ view: UIView) {
-        if let url = copiedURL,
+        if let url = TodayModel.copiedURL,
             let encodedString = url.absoluteString.escape() {
             openContainingApp("?url=\(encodedString)")
         }
-    }
-}
-
-extension UIButton {
-    func setBackgroundColor(_ color: UIColor, forState state: UIControl.State) {
-        let colorView = UIView(frame: CGRect(width: 1, height: 1))
-        colorView.backgroundColor = color
-
-        UIGraphicsBeginImageContext(colorView.bounds.size)
-        if let context = UIGraphicsGetCurrentContext() {
-            colorView.layer.render(in: context)
-        }
-        let colorImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-
-        self.setBackgroundImage(colorImage, for: state)
-    }
-}
-
-class ImageButtonWithLabel: UIView {
-
-    lazy var button = UIButton()
-    lazy var label = UILabel()
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        performLayout()
-    }
-
-    func performLayout() {
-        addSubview(button)
-        addSubview(label)
-        button.imageView?.contentMode = .scaleAspectFit
-
-        button.snp.makeConstraints { make in
-            make.centerX.equalTo(self)
-            make.top.equalTo(self.safeAreaLayoutGuide).offset(5)
-            make.right.greaterThanOrEqualTo(self.safeAreaLayoutGuide).offset(40)
-            make.left.greaterThanOrEqualTo(self.safeAreaLayoutGuide).inset(40)
-            make.height.greaterThanOrEqualTo(60)
-        }
-
-        label.snp.makeConstraints { make in
-            make.top.equalTo(button.snp.bottom).offset(10)
-            make.leading.trailing.bottom.equalTo(self)
-            make.height.equalTo(10)
-        }
-
-        label.numberOfLines = 1
-        label.lineBreakMode = .byWordWrapping
-        label.textAlignment = .center
-    }
-
-    func addTarget(_ target: AnyObject?, action: Selector, forControlEvents events: UIControl.Event) {
-        button.addTarget(target, action: action, for: events)
-    }
-}
-
-class ButtonWithSublabel: UIButton {
-    lazy var subtitleLabel = UILabel()
-    lazy var label = UILabel()
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    convenience init() {
-        self.init(frame: .zero)
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        performLayout()
-    }
-
-    fileprivate func performLayout() {
-        let titleLabel = self.label
-        self.titleLabel?.removeFromSuperview()
-        addSubview(titleLabel)
-
-        let imageView = self.imageView!
-        let subtitleLabel = self.subtitleLabel
-        self.addSubview(subtitleLabel)
-
-        imageView.snp.makeConstraints { make in
-            make.centerY.left.equalTo(10)
-            make.width.equalTo(TodayUX.copyLinkImageWidth)
-        }
-
-        titleLabel.snp.makeConstraints { make in
-            make.left.equalTo(imageView.snp.right).offset(10)
-            make.trailing.top.equalTo(self)
-            make.height.greaterThanOrEqualTo(12)
-        }
-
-        subtitleLabel.lineBreakMode = .byTruncatingTail
-        subtitleLabel.snp.makeConstraints { make in
-            make.bottom.equalTo(self).inset(10)
-            make.top.equalTo(titleLabel.snp.bottom)
-            make.leading.trailing.equalTo(titleLabel)
-            make.height.greaterThanOrEqualTo(10)
-        }
-    }
-
-    override func setTitle(_ text: String?, for state: UIControl.State) {
-        self.label.text = text
-        super.setTitle(text, for: state)
     }
 }
